cloud config templates: add migration script
Add migration script to migrate cloud-config templates from ConfigMaps
to Secrets.

To reduce the number differend kinds of Kubernetes objects that the UI
needs to query and cache and to make the API more homogenous, the
cloud-config templates should be stored in Secrets rather than
ConfigMaps. This means that users who upgrade need to have their
existing cloud-config templates migrated when they update.
This change adds a script which can be executed during an upgrade, that
will take care of this migration automatically so users don't lose
templates that have been grandfathered in from before this migration.

Related-to: https://github.com/harvester/harvester/issues/2329

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>